requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
feedparser==6.0.10
tweepy==4.14.0
python-dateutil==2.8.2
